move global_configuration functions to their own file
